Fix incorrect install location of networking default file
Ticket: none
Reviewed By: roopa,satish
Testing Done: built, installed, verified correct location

The install file resulted in creating /etc/default/networking/ as a
directory and installing the networking.default file in it.  Rename
networking.default to networking, and change the install file.
Normally dh_installinit handles all this, but since the package is
ifupdown2, and the file is networking, that doesn't happen.

Also, because we had created a directory with the name of what we want
to install as a file, we need to remove it if it exists.  This addition
of a preinst file should not go upstream, and should be removed in a few
weeks when everybody has re-installed enough.