ifupdownmain: warn user if an interface matches a regex but doesn't exists in the config anymore
Ticket: CM-13316
Reviewed By: Roopa, Nikhil G
Testing Done:

auto v0
iface v0
      vlan-id 100
      vlan-raw-device tap1

auto v1
iface v1
      vlan-id 100
      vlan-raw-device tap2

auto v2
iface v2
      vlan-id 100
      vlan-raw-device tap3

auto br0
iface br0
        bridge-ports tap9 regex v.*

In this specific case, if you remove v0 then execute ifreload ifupdown2 will
barf and say that v0 still exists as a dependency of br0.

From roopa:
The problem here is that the dependency gathering should have been done after
the ifdown's. However the ifdown's  need the dependency tree for built in
interfaces which could be part of globs. And this dependency tree needs to be
the current state of things so can't do the selective picking of interfaces
either. So, its a chicken and egg problem.

Ideally when one uses regex, if he is removing an interface that was matched by
a regex, he has to do a 'ifdown' manually today. 'ifdown v0' in this case.

So today we decided to change the warning message to let the user know that he
should manually down the interface that matches a regex.
This is not a permanent fix but a work-around. We don't want to introduce new
changes that may introduce false positive case for existing and working config

post 3.2:
-----------
we should consider doing the down's in two phases during ifreload
a) down everything that we are very certain was removed from the file
(to make sure regex's don't catch it).
b) build dependency tree and run through the down's once more, to
catch all glob matches for built-in interfaces

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>